Laboratory/Lr3/celery/requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
python-dotenv
celery
redis
requests
bs4
